requirements.txt
opencv-python-headless
numpy==1.24.2
mediapipe
tflite-runtime
python-telegram-bot==13.15
gpiozero
picamera2
tensorflow-aarch64==2.16.1
insightface
onnxruntime
imutils
#scipy cài bằng apt không cài bằng pip
PySide6
fastapi
uvicorn
